requirements.txt
Django==4.2.7
django-environ==0.11.2
django-crispy-forms==2.1
crispy-tailwind==0.5.0
psycopg2-binary==2.9.9
tensorflow==2.14.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
django-rest-framework==0.1.0
python-instagram==1.3.2
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.0
python-dateutil==2.8.2